// server/src/modules/employees/repository/employee.types.ts
export interface EmployeeDB {
  MGMT_PASS: string;
  ID: number;
  NAME: string;
  ID_SECTOR: number;
  DEPARTMENT: string;
  STATUS: number;
  LOCATION: string;
  FIRSTNAME: string;
  LASTNAME: string;
  MIDDLENAME: string;
  BANK_CARD: string;
  CARD_HOLDER: string;
  PHONE: string;
  PERMISSION_GROUP_ID: number;
}